Added Debug Progress for segmentation/ Added training data with proper keys
I was unable to figure out segmentation, however I did make progress past two major errors. At this point, the segmentID is drawing from the label vector that is stored in the .h5 data, and the label_vector key does not seem to store the relevant data. We will fix tomorrow.